Test verbose output
