+++ semgrep/tests/e2e/snapshots/test_cli_test/test_cli_test_verbose/results.json
@@ -1,0 +1,17 @@
+targets/cli_test/basic/basic.py
+rule:rules.cli_test.basic.basic-test: Basic test
+2:print(1 == 1)
+
+============================[ summary ]============================
+Total time: x.xxx Config time: x.xxx Core time: x.xxx Ignores time: x.xxx
+
+Semgrep-core time:
+Total CPU time: x.xxx  File parse time: x.xxx  Rule parse time: x.xxx  Match time: x.xxx
+Slowest x.xxx files
+targets/cli_test/basic/basic.py                                        ( 35B):  x.xxx
+Slowest 5 rules to run (excluding parse time)
+rules.cli_test.basic.basic-test:                                        run time x.xxx  match time x.xxx
+
+Analyzed: 1 python files ( 35B in x.xxx seconds)
+Errors:   0 files with errors
+
